remove the timer thread
in the implementation from mio-extras, the timers are stored in a hashed
timing wheel, and a separate thread is used to send events to mio when
the next timer triggers.
The thread sleeps uses an atomic int to know for how much time it needs
to sleep, and that atomic int is updated when a new timer is added, and
that timer's expiration date is closer than the current next date.
If that timer cancels, the next date is not updated though.

Unfortunately, sozu uses timers to implement session timeouts, and for
sessions that worked properly, those timeouts will be canceled. The
result is that the timer thread wakes up too frequently and consumes too
much CPU.

This commit removes the timer thread, and instead adds a method to know
when the next timer will trigger, use it as timeout for the poll() call,
and use it to know when we can check the timer. This results in very
cheap timer usage